SL-1551 refactored packages
