+++ src/main/java/uk/gov/hmcts/reform/sandl/snlevents/controllers/HearingController.java
@@ -15,7 +15,7 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 import uk.gov.hmcts.reform.sandl.snlevents.model.db.UserTransaction;
 import uk.gov.hmcts.reform.sandl.snlevents.model.request.HearingSessionRelationship;
-import uk.gov.hmcts.reform.sandl.snlevents.model.request.search.SearchCriteria;
+import uk.gov.hmcts.reform.sandl.snlevents.model.specifications.SearchCriteria;
 import uk.gov.hmcts.reform.sandl.snlevents.model.response.HearingInfo;
 import uk.gov.hmcts.reform.sandl.snlevents.repository.db.HearingRepository;
 import uk.gov.hmcts.reform.sandl.snlevents.service.HearingPartService;
