+++ src/main/java/uk/gov/hmcts/reform/sandl/snlevents/service/HearingService.java
@@ -1,18 +1,16 @@
 package uk.gov.hmcts.reform.sandl.snlevents.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.sandl.snlevents.model.db.Hearing;
-import uk.gov.hmcts.reform.sandl.snlevents.model.request.search.HearingSpecificationBuilder;
-import uk.gov.hmcts.reform.sandl.snlevents.model.request.search.SearchCriteria;
+import uk.gov.hmcts.reform.sandl.snlevents.model.specifications.HearingSpecificationBuilder;
+import uk.gov.hmcts.reform.sandl.snlevents.model.specifications.SearchCriteria;
 import uk.gov.hmcts.reform.sandl.snlevents.model.response.HearingInfo;
 import uk.gov.hmcts.reform.sandl.snlevents.repository.db.HearingRepository;
 
 import java.util.List;
-import java.util.stream.Collectors;
 import javax.persistence.EntityManager;
 
 @Service
